test: add testRule with initial rule unit tests (#26)
## PR Checklist

- [x] Addresses an existing open issue: fixes #23
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/octoguide/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/octoguide/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

Adds a `testRule` helper that takes in a rule, a deep partial of an
entity, and a deep partial of testing context. Those deep partials are
set up with `Proxy`s to error if an unknown property is accessed.

🗺️